refactor: Line에서 points를 검증하지 않는다
* 완전한 인자들을 제공받아야 한다.
