+++ src/main/java/laddergame/domain/Line.java
@@ -4,30 +4,17 @@ import static laddergame.domain.Point.of;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.IntStream;
 
 public class Line {
     private final List<Point> points = new ArrayList<>();
 
-    public Line(final List<Boolean> rungExistsAtPoint) {
-        validate(rungExistsAtPoint);
-        for (final Boolean isFilled : rungExistsAtPoint) {
-            points.add(of(isFilled));
-        }
-    }
-
-    private void validate(final List<Boolean> points) {
+    public Line(final List<Boolean> points) {
         if (points.isEmpty()) {
             throw new IllegalArgumentException("사다리의 가로는 한 칸 이상이어야합니다.");
         }
-        if (hasAdjacentRung(points)) {
-            throw new IllegalArgumentException("사다리의 가로 라인은 겹칠 수 없습니다.");
+        for (final Boolean isFilled : points) {
+            this.points.add(of(isFilled));
         }
-    }
-
-    private static boolean hasAdjacentRung(List<Boolean> rungExistsAtColumn) {
-        return IntStream.range(0, rungExistsAtColumn.size() - 1)
-                .anyMatch(pointIndex -> rungExistsAtColumn.get(pointIndex) && rungExistsAtColumn.get(pointIndex + 1));
     }
 
     public List<Point> getPoints() {
